# specs/002-rag-chatbot/contracts/openapi.yaml
openapi: 3.0.0
info:
  title: RAG Chatbot API
  description: API for the RAG chatbot system that answers questions based on book content
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.example.com
    description: Production server

paths:
  /chat:
    post:
      summary: General chat with the RAG system
      description: Send a query to the RAG system and receive a response based on book content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Successful response from the chatbot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - api_key: []

  /chat/selected-text:
    post:
      summary: Chat about selected text
      description: Ask questions specifically about selected text on the book page
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SelectedTextRequest'
      responses:
        '200':
          description: Successful response focused on the selected text
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - api_key: []

components:
  schemas:
    ChatRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          description: The user's question
          example: "Can you explain the concept of forward kinematics?"
        page:
          type: string
          description: The current page identifier (optional)
          example: "chapter3/section2"
        chapter:
          type: string
          description: The current chapter (optional, used for context)
          example: "Robot Kinematics"
        selected_text:
          type: string
          description: Text selected by the user (optional)
          example: "Forward kinematics is the process of determining the position and orientation of the end-effector based on the joint angles."

    SelectedTextRequest:
      type: object
      required:
        - query
        - selected_text
      properties:
        query:
          type: string
          description: The user's question about the selected text
          example: "What does this mean?"
        selected_text:
          type: string
          description: The text the user has selected
          example: "Forward kinematics is the process of determining the position and orientation of the end-effector based on the joint angles."
        page:
          type: string
          description: The current page identifier (optional)
          example: "chapter3/section2"
        chapter:
          type: string
          description: The current chapter (optional, used for context)
          example: "Robot Kinematics"

    ChatResponse:
      type: object
      properties:
        response:
          type: string
          description: The AI-generated response
          example: "Forward kinematics refers to the mathematical process of determining the position and orientation of a robot's end-effector based on the known joint angles."
        sources:
          type: array
          items:
            $ref: '#/components/schemas/Source'
          description: List of sources used in generating the response
        confidence:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          description: Confidence level of the response
          example: 0.92

    Source:
      type: object
      properties:
        content_id:
          type: string
          description: Unique identifier for the content chunk
          example: "ch3-sec2-para4"
        chapter:
          type: string
          description: Chapter where the content appears
          example: "Robot Kinematics"
        section:
          type: string
          description: Section where the content appears
          example: "Forward Kinematics"

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Query too long"
        code:
          type: string
          description: Error code
          example: "QUERY_TOO_LONG"

  securitySchemes:
    api_key:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for rate limiting and authentication